# api.yml
...
openapi: 3.0.1
info:
  title: GeoSupport_ms
  description: |-
    Hello world, and welcome to the GeoSupport_ms API gide!
    GeoSupport is an ASP.Net Core microservice for the GeoSmart project.
    Go ahead and check all the request you can use :) .
  version: v1
paths:
...
  "/api/Flag":
    get:
      tags:
      - Flag
      summary: Gets a list of Flag elements
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Flag"
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Flag"
            text/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Flag"
    post:
      tags:
      - Flag
      summary: Inserts a new Flag register in database
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Flag"
          text/json:
            schema:
              "$ref": "#/components/schemas/Flag"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/Flag"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Flag"
            application/json:
              schema:
                "$ref": "#/components/schemas/Flag"
            text/json:
              schema:
                "$ref": "#/components/schemas/Flag"
  "/api/Flag/{id}":
    get:
      tags:
      - Flag
      summary: Gets a specific Flag element given its Id
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/FlagColors"
            application/json:
              schema:
                "$ref": "#/components/schemas/FlagColors"
            text/json:
              schema:
                "$ref": "#/components/schemas/FlagColors"
        '404':
          description: If there is not item whit that id
    put:
      tags:
      - Flag
      summary: Updates a specific Flag element given its Id
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Flag"
          text/json:
            schema:
              "$ref": "#/components/schemas/Flag"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/Flag"
      responses:
        '200':
          description: Success
        '400':
          description: If given id not valid
        '404':
          description: When an error is producted updating db
    delete:
      tags:
      - Flag
      summary: Deletes a specific Flag register given its Id
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/Flag"
            application/json:
              schema:
                "$ref": "#/components/schemas/Flag"
            text/json:
              schema:
                "$ref": "#/components/schemas/Flag"
        '404':
          description: If there is not item whit that id
...
components:
  schemas:
    Flag:
      required:
      - flagImage
      type: object
      properties:
        id_flag:
          type: integer
          format: int32
        flagImage:
          maxLength: 2000
          type: string
        colors:
          type: array
          items:
            "$ref": "#/components/schemas/Color"
          nullable: true
      additionalProperties: false
    Color:
      required:
      - hexCode
      - name
      type: object
      properties:
        id_color:
          type: integer
          format: int32
        name:
          maxLength: 20
          type: string
        hexCode:
          maxLength: 20
          type: string
        flags:
          type: array
          items:
            "$ref": "#/components/schemas/Flag"
          nullable: true
      additionalProperties: false
...
    ColorOrder:
      type: object
      properties:
        id_color:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        hexCode:
          type: string
          nullable: true
        order:
          type: integer
          format: int32
      additionalProperties: false
    FlagColors:
      type: object
      properties:
        id_flag:
          type: integer
          format: int32
        flagImage:
          type: string
          nullable: true
        colors:
          type: array
          items:
            "$ref": "#/components/schemas/ColorOrder"
          nullable: true
      additionalProperties: false